+++ build-config.ps1
@@ -20,6 +20,7 @@ $SPLICE_MODIFIED = @(
     "graphics\LevelPanel.java",
     "graphics\MenuBar.java",
     "graphics\GamePanel.java",
+    "io\ErrorLog.java",
     "io\LevelFactory.java",
     "io\SuccPaths.java",
     "io\TWSWriter.java"